# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is Swagger Documentation API For BackEnd DBO Test.",
        "title": "DBO BackEnd Test API",
        "contact": {
            "name": "Afdhalul Ihsan",
            "email": "afdhalulhsn74@gmail.com"
        },
        "license": {
            "name": "Apache 2.0"
        },
        "version": "1.0"
    },
    "host": "localhost:8830",
    "basePath": "/",
    "paths": {
...
        "/admin/produk/add": {
            "post": {
                "description": "get the status of server.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AddProduct"
                ],
                "summary": "Show the status of server.",
                "parameters": [
                    {
                        "description": "AddProduct",
                        "name": "Payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AddProduct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseData"
                        }
                    }
                }
            }
        },
...
        "/api/v1/order/add": {
            "post": {
                "description": "get the status of server.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "InsertOrder"
                ],
                "summary": "Show the status of server.",
                "parameters": [
                    {
                        "description": "InsertOrder",
                        "name": "Payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.InsertOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/v1/order/detail/:id": {
            "get": {
                "description": "get the status of server.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GetDetailOrder"
                ],
                "summary": "Show the status of server.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.DataOrder"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "get the status of server.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DeleteOrder"
                ],
                "summary": "Show the status of server.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/v1/order/paging": {
            "get": {
                "description": "get the status of server.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GetAllOrderPagination"
                ],
                "summary": "Show the status of server.",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.DataOrder"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/order/search": {
            "get": {
                "description": "get the status of server.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "SearchAllOrder"
                ],
                "summary": "Show the status of server.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "key",
                        "name": "key",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.DataOrder"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/order/update": {
            "put": {
                "description": "get the status of server.",
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UpdateOrder"
                ],
                "summary": "Show the status of server.",
                "parameters": [
                    {
                        "description": "UpdateOrder",
                        "name": "Payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.ResponseData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entity.CustomerData": {
            "type": "object",
            "properties": {
                "dateOfBirth": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "noHp": {
                    "type": "string"
                },
                "noId": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "entity.DataOrder": {
            "type": "object",
            "properties": {
                "customerData": {
                    "$ref": "#/definitions/entity.CustomerData"
                },
                "id": {
                    "type": "string"
                },
                "idCustomer": {
                    "type": "string"
                },
                "idProduct": {
                    "type": "string"
                },
                "orderDate": {
                    "type": "string"
                },
                "productData": {
                    "$ref": "#/definitions/entity.ProductData"
                },
                "qty": {
                    "type": "integer"
                }
            }
        },
...
        "entity.ProductData": {
            "type": "object",
            "properties": {
                "harga": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "kodeProduk": {
                    "type": "string"
                },
                "namaProduck": {
                    "type": "string"
                }
            }
        },
...
        "model.AddProduct": {
            "type": "object",
            "properties": {
                "harga": {
                    "type": "integer"
                },
                "kodeProduk": {
                    "type": "string"
                },
                "namaProduck": {
                    "type": "string"
                }
            }
        },
...
        "model.InsertOrder": {
            "type": "object",
            "properties": {
                "id_customer": {
                    "type": "string"
                },
                "id_product": {
                    "type": "string"
                },
                "qty": {
                    "type": "integer"
                }
            }
        },
...
        "model.ResponseData": {
            "type": "object",
            "properties": {
                "data": {},
                "error_code": {
                    "type": "string"
                },
                "is_error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "model.UpdateOrder": {
            "type": "object",
            "properties": {
                "id_order": {
                    "type": "string"
                },
                "qty": {
                    "type": "integer"
                }
            }
        }
    }
}